refactor: reorganize folder structure and update import paths (#4)
