+++ src/components/MainBanner/MainBanner.jsx
