+++ src/components/MainBanner/MainBanner.styled.js
